+++ NEWS.md
@@ -1,5 +1,12 @@
 # PlotR app development version
 
+## PlotR 22.12.1
+
+### Updates
+- added placeholders to several select inputs as long as data is missing (#14)
+- rename tab "File Upload(s)" to "File Import(s)" for more consistency with the placeholders
+- tab "Style plot": add numeric inputs to sliders for the plot ranges (#13)
+
 ## PlotR 22.11.1
 
 ### New Features